stub services ready: wallet, top_up, money, withdraw, supernode

+++ m2m-wallet/pkg/services/withdraw/withdraw_data.go
@@ -1,7 +1,7 @@
 package withdraw
 
 var ctxWithdraw = struct {
-	withdrawFee float64
+	withdrawFee float32
 }{
 	withdrawFee: 200,
 }
